Tidying up final changes

+++ src/main/java/com/example/lockly/controllerLayer/dtos/passwords/PasswordWithoutFolderDto.java
@@ -4,5 +4,5 @@ import com.example.lockly.controllerLayer.dtos.UserDto;
 import lombok.Builder;
 
 @Builder
-public record PasswordWithoutFolderDto(Long id, String name, String content, UserDto userDto) {
+public record PasswordWithoutFolderDto(Integer id, String name, String content, UserDto userDto) {
 }
